requirements.txt
langchain
langchain-core
langchain-groq
langsmith
pydantic
python-dotenv